+++ opentelemetry-otlp/tests/integration_test/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod logs_asserter;
-pub mod metrics_asserter;
+pub mod metric_helpers;
 pub mod test_utils;
 pub mod trace_asserter;
